+++ src/test/java/ec/com/jmgorduez/ACMEEmployeesPayment/dataGenerator/TestDataGenerator.java
@@ -6,9 +6,9 @@ public class TestDataGenerator {
 
     public static final Integer ONE = 1;
 
-    public static final Float _45_USD = 45F;
-    public static final Float _50_USD = 50F;
-    public static final Float _87_USD_50_c = 87.5F;
+    public static final Double _45_USD = 45D;
+    public static final Double _50_USD = 50D;
+    public static final Double _87_USD_50_c = 87.5D;
 
     public static final LocalTime _09_01 = LocalTime.parse("09:01");
     public static final LocalTime _12_01 = LocalTime.parse("12:01");
